move organization admin
Moved the organization admin components from the admin page to the organizations page under the gear icon. Path is /organizations/list/admin
