+++ frontend/src/app/admin/admin-routing.module.ts
@@ -4,7 +4,6 @@ import { AdminComponent } from './admin.component';
 import { AdminRoleDetailsComponent } from './roles/details/admin-role-details.component';
 import { AdminRolesListComponent } from './roles/list/admin-roles-list.component';
 import { AdminUsersListComponent } from './users/list/admin-users-list.component';
-import { AdminOrganizationListComponent } from './organization/list/admin-organization-list.component';
 
 const routes: Routes = [
   {
@@ -13,8 +12,7 @@ const routes: Routes = [
     children: [
       AdminUsersListComponent.Route,
       AdminRolesListComponent.Route,
-      AdminRoleDetailsComponent.Route,
-      AdminOrganizationListComponent.Route
+      AdminRoleDetailsComponent.Route
     ]
   }
 ];
