+++ frontend/src/app/admin/admin.component.ts
@@ -12,8 +12,7 @@ export class AdminComponent {
 
   public links = [
     { label: 'Users', path: '/admin/users' },
-    { label: 'Roles', path: '/admin/roles' },
-    { label: 'Organizations', path: '/admin/organizations' }
+    { label: 'Roles', path: '/admin/roles' }
   ];
 
   constructor(public profileService: ProfileService) {
