+++ frontend/src/app/admin/admin.module.ts
@@ -19,15 +19,13 @@ import { ReactiveFormsModule } from '@angular/forms';
 import { AdminUsersListComponent } from './users/list/admin-users-list.component';
 import { AdminRolesListComponent } from './roles/list/admin-roles-list.component';
 import { AdminRoleDetailsComponent } from './roles/details/admin-role-details.component';
-import { AdminOrganizationListComponent } from './organization/list/admin-organization-list.component';
 
 @NgModule({
   declarations: [
     AdminComponent,
     AdminUsersListComponent,
     AdminRolesListComponent,
-    AdminRoleDetailsComponent,
-    AdminOrganizationListComponent
+    AdminRoleDetailsComponent
   ],
   imports: [
     CommonModule,
