--- frontend/src/app/admin/organization/list/admin-organization-list.component.css
@@ -1,30 +1,0 @@
-/** 
-* admin-organization-list.component.css 
-* 
-* The admin organization list page should provide
-* a simple, easily readable form for users to view
-* all organizations.
-*
-*/
-
-.mat-mdc-row .mat-mdc-cell {
-    border-bottom: 1px solid transparent;
-    border-top: 1px solid transparent;
-    cursor: pointer;
-}
-
-.mat-mdc-row:hover .mat-mdc-cell {
-    border-color: white;
-}
-
-.header {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-}
-
-.row {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-}
