--- frontend/src/app/admin/organization/list/admin-organization-list.component.html
@@ -1,26 +1,0 @@
-<!-- HTML Structure of Admin Organizations List -->
-<div class="content" *ngIf="organizations$ | async as organizations">
-  <table mat-table [dataSource]="organizations">
-    <ng-container matColumnDef="name">
-      <th mat-header-cell *matHeaderCellDef>
-        <div class="header">
-          Organizations
-          <button mat-stroked-button (click)="createOrganization()">
-            Create
-          </button>
-        </div>
-      </th>
-      <td mat-cell *matCellDef="let element">
-        <div class="row">
-          <p>{{ element.name }}</p>
-          <button mat-stroked-button (click)="deleteOrganization(element)">
-            Delete
-          </button>
-        </div>
-      </td>
-    </ng-container>
-
-    <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
-    <tr mat-row *matRowDef="let row; columns: displayedColumns"></tr>
-  </table>
-</div>
