+++ frontend/src/app/organization/organization-page/organization-page.component.ts
@@ -63,7 +63,7 @@ export class OrganizationPageComponent implements OnInit {
       'organizations.*',
       '*',
       '',
-      'admin/organizations'
+      'organizations/list/admin'
     );
   }
 }
