+++ frontend/src/app/organization/organization-routing.module.ts
@@ -12,11 +12,13 @@ import { RouterModule, Routes } from '@angular/router';
 import { OrganizationPageComponent } from './organization-page/organization-page.component';
 import { OrganizationDetailsComponent } from './organization-details/organization-details.component';
 import { OrganizationEditorComponent } from './organization-editor/organization-editor.component';
+import { OrganizationListAdminComponent } from './organization-admin/list/organization-list-admin.component';
 
 const routes: Routes = [
   OrganizationPageComponent.Route,
   OrganizationDetailsComponent.Route,
-  OrganizationEditorComponent.Route
+  OrganizationEditorComponent.Route,
+  OrganizationListAdminComponent.Route
 ];
 
 @NgModule({
